Rename function_base import
