+++ numpy/lib/polynomial.py
@@ -18,7 +18,7 @@ from numpy.core import (isscalar, abs, finfo, atleast_1d, hstack, dot, array,
 from numpy.core import overrides
 from numpy.exceptions import RankWarning
 from numpy.lib.twodim_base import diag, vander
-from numpy.lib.function_base import trim_zeros
+from numpy.lib._function_base_impl import trim_zeros
 from numpy.lib.type_check import iscomplex, real, imag, mintypecode
 from numpy.linalg import eigvals, lstsq, inv
