+++ numpy/ma/core.pyi
@@ -12,10 +12,7 @@ from numpy import (
     ones_like as ones_like,
     squeeze as squeeze,
     zeros_like as zeros_like,
-)
-
-from numpy.lib.function_base import (
-    angle as angle,
+    angle as angle
 )
 
 # TODO: Set the `bound` to something more suitable once we
